lua: make more results visible
Telescope is currently limiting the number of results to 250 without
an official way of configuring this limit externally.

This new behavior prevents us from seeing all of the grep results when
more than 250 results are found.

Workaround this behavior by overriding the undocumented internal limit value
until the issue is resolved in #nvim-telescope/telescope.nvim/2779.